Pull out PartnerAlertType into seperate file
