+++ src/schema/v2/Alerts/index.ts
@@ -440,13 +440,6 @@ export const AlertsConnectionType = connectionWithCursorInfo({
   edgeFields: AlertsEdgeFields,
 }).connectionType
 
-export const PartnerAlertsSummaryFields = {
-  id: {
-    type: AlertType,
-    resolve: ({ id }) => id,
-  },
-}
-
 export const AlertsSummaryFields = {
   topHit: {
     type: AlertType,
